Move job_table to IsDatabase
