+++ pyiron_base/database/generic.py
@@ -13,6 +13,7 @@ import time
 import os
 from datetime import datetime
 from pyiron_base.generic.util import deprecate
+import pandas
 from sqlalchemy import (
     create_engine,
     MetaData,
@@ -65,6 +66,53 @@ class IsDatabase(ABC):
     def viewer_mode(self):
         return self.view_mode
     viewer_mode.__doc__ = view_mode.__doc__
+
+    def job_table(
+            self,
+            sql_query,
+            user,
+            project_path,
+            recursive=True,
+            columns=None,
+            all_columns=False,
+            sort_by="id",
+            max_colwidth=200,
+            full_table=False,
+            element_lst=None,
+            job_name_contains='',
+    ):
+        """
+        Access the job_table.
+
+        Args:
+            sql_query (str): SQL query to enter a more specific request
+            user (str): username of the user whoes user space should be searched
+            project_path (str): root_path - this is in contrast to the project_path in GenericPath
+            recursive (bool): search subprojects [True/False]
+            columns (list): by default only the columns ['job', 'project', 'chemicalformula'] are selected, but the
+                            user can select a subset of ['id', 'status', 'chemicalformula', 'job', 'subjob', 'project',
+                            'projectpath', 'timestart', 'timestop', 'totalcputime', 'computer', 'hamilton', 'hamversion',
+                            'parentid', 'masterid']
+            all_columns (bool): Select all columns - this overwrites the columns option.
+            sort_by (str): Sort by a specific column
+            max_colwidth (int): set the column width
+            full_table (bool): Whether to show the entire pandas table
+            element_lst (list): list of elements required in the chemical formular - by default None
+            job_name_contains (str): a string which should be contained in every job_name
+
+        Returns:
+            pandas.Dataframe: Return the result as a pandas.Dataframe object
+        """
+        return self._get_job_table(
+                project=project_path,
+                recursive=recursive,
+                columns=columns,
+                all_columns=all_columns,
+                sort_by=sort_by,
+                max_colwidth=max_colwidth,
+                full_table=full_table,
+                job_name_contains=job_name_contains
+        )
 
 class ConnectionWatchDog(Thread):
     """
@@ -135,7 +183,6 @@ class ConnectionWatchDog(Thread):
         """
         self._queue.put(False)
         self.join()
-
 
 class AutorestoredConnection:
     def __init__(self, engine, timeout=60):
@@ -251,6 +298,152 @@ class DatabaseAccess(IsDatabase):
     @deprecate("use view_mode")
     def viewer_mode(self, value):
         self.view_mode = value
+
+    def _job_dict(
+        self,
+        sql_query,
+        user,
+        project_path,
+        recursive,
+        job=None,
+        sub_job_name="%",
+        element_lst=None,
+    ):
+        """
+        Internal function to access the database from the project directly.
+
+        Args:
+            sql_query (str): SQL query to enter a more specific request
+            user (str): username of the user whoes user space should be searched
+            project_path (str): root_path - this is in contrast to the project_path in GenericPath
+            recursive (bool): search subprojects [True/False]
+            job (str): job_name - by default None
+            sub_job_name (str): path inside the HDF5 file - "%" by default to accept any path
+            element_lst (list): list of elements required in the chemical formular - by default None
+
+        Returns:
+            list: the function returns a list of dicts like get_items_sql, but it does not format datetime:
+                 [{'chemicalformula': u'Ni108',
+                  'computer': u'mapc157',
+                  'hamilton': u'LAMMPS',
+                  'hamversion': u'1.1',
+                  'id': 24,
+                  'job': u'DOF_1_0',
+                  'parentid': 21L,
+                  'project': u'lammps.phonons.Ni_fcc',
+                  'projectpath': u'D:/PyIron/PyIron_data/projects',
+                  'status': u'finished',
+                  'timestart': datetime.datetime(2016, 6, 24, 10, 17, 3, 140000),
+                  'timestop': datetime.datetime(2016, 6, 24, 10, 17, 3, 173000),
+                  'totalcputime': 0.033,
+                  'username': u'test'},
+                 {'chemicalformula': u'Ni108',
+                  'computer': u'mapc157',
+                  'hamilton': u'LAMMPS',
+                  'hamversion': u'1.1',
+                  'id': 21,
+                  'job': u'ref',
+                  'parentid': 20L,
+                  'project': u'lammps.phonons.Ni_fcc',
+                  'projectpath': u'D:/PyIron/PyIron_data/projects',
+                  'status': u'finished',
+                  'timestart': datetime.datetime(2016, 6, 24, 10, 17, 2, 429000),
+                  'timestop': datetime.datetime(2016, 6, 24, 10, 17, 2, 463000),
+                  'totalcputime': 0.034,
+                  'username': u'test'},.......]
+
+        """
+        dict_clause = {}
+        # FOR GET_ITEMS_SQL: clause = []
+        if user is not None:
+            dict_clause["username"] = str(user)
+            # FOR GET_ITEMS_SQL: clause.append("username = '" + self.user + "'")
+        if sql_query is not None:
+            # FOR GET_ITEMS_SQL: clause.append(self.sql_query)
+            if "AND" in sql_query:
+                cl_split = sql_query.split(" AND ")
+            elif "and" in sql_query:
+                cl_split = sql_query.split(" and ")
+            else:
+                cl_split = [sql_query]
+            dict_clause.update(
+                {str(element.split()[0]): element.split()[2] for element in cl_split}
+            )
+        if job is not None:
+            dict_clause["job"] = str(job)
+
+        if project_path == "./":
+            project_path = ""
+        if recursive:
+            dict_clause["project"] = str(project_path) + "%"
+        else:
+            dict_clause["project"] = str(project_path)
+        if sub_job_name is None:
+            dict_clause["subjob"] = None
+        elif sub_job_name != "%":
+            dict_clause["subjob"] = str(sub_job_name)
+        if element_lst is not None:
+            dict_clause["element_lst"] = element_lst
+
+        s.logger.debug("sql_query: %s", str(dict_clause))
+        return self.get_items_dict(dict_clause)
+
+    def _get_job_table(
+            self,
+            sql_query,
+            user,
+            project_path,
+            recursive=True,
+            columns=None,
+            all_columns=False,
+            sort_by="id",
+            max_colwidth=200,
+            full_table=False,
+            element_lst=None,
+            job_name_contains='',
+    ):
+        if columns is None:
+            columns = ["job", "project", "chemicalformula"]
+        all_db = [
+            "id",
+            "status",
+            "chemicalformula",
+            "job",
+            "subjob",
+            "projectpath",
+            "project",
+            "timestart",
+            "timestop",
+            "totalcputime",
+            "computer",
+            "hamilton",
+            "hamversion",
+            "parentid",
+            "masterid",
+        ]
+
+        if all_columns:
+            columns = all_db
+        job_dict = self._job_dict(
+            sql_query=sql_query,
+            user=user,
+            project_path=project_path,
+            recursive=recursive,
+            element_lst=element_lst,
+        )
+        if full_table:
+            pandas.set_option('display.max_rows', None)
+            pandas.set_option('display.max_columns', None)
+        else:
+            pandas.reset_option('display.max_rows')
+            pandas.reset_option('display.max_columns')
+        pandas.set_option("display.max_colwidth", max_colwidth)
+        df = pandas.DataFrame(job_dict, columns=columns)
+        if job_name_contains != '':
+            df = df[df.job.str.contains(job_name_contains)]
+        if sort_by in columns:
+            return df[columns].sort_values(by=sort_by)
+        return df[columns]
 
     # Internal functions
     def __del__(self):
