+++ pyiron_base/database/jobtable.py
@@ -24,96 +24,6 @@ __date__ = "Sep 1, 2017"
 s = Settings()
 
 
-def _job_dict(
-    database,
-    sql_query,
-    user,
-    project_path,
-    recursive,
-    job=None,
-    sub_job_name="%",
-    element_lst=None,
-):
-    """
-    Internal function to access the database from the project directly.
-
-    Args:
-        database (DatabaseAccess): Database object
-        sql_query (str): SQL query to enter a more specific request
-        user (str): username of the user whoes user space should be searched
-        project_path (str): root_path - this is in contrast to the project_path in GenericPath
-        recursive (bool): search subprojects [True/False]
-        job (str): job_name - by default None
-        sub_job_name (str): path inside the HDF5 file - "%" by default to accept any path
-        element_lst (list): list of elements required in the chemical formular - by default None
-
-    Returns:
-        list: the function returns a list of dicts like get_items_sql, but it does not format datetime:
-             [{'chemicalformula': u'Ni108',
-              'computer': u'mapc157',
-              'hamilton': u'LAMMPS',
-              'hamversion': u'1.1',
-              'id': 24,
-              'job': u'DOF_1_0',
-              'parentid': 21L,
-              'project': u'lammps.phonons.Ni_fcc',
-              'projectpath': u'D:/PyIron/PyIron_data/projects',
-              'status': u'finished',
-              'timestart': datetime.datetime(2016, 6, 24, 10, 17, 3, 140000),
-              'timestop': datetime.datetime(2016, 6, 24, 10, 17, 3, 173000),
-              'totalcputime': 0.033,
-              'username': u'test'},
-             {'chemicalformula': u'Ni108',
-              'computer': u'mapc157',
-              'hamilton': u'LAMMPS',
-              'hamversion': u'1.1',
-              'id': 21,
-              'job': u'ref',
-              'parentid': 20L,
-              'project': u'lammps.phonons.Ni_fcc',
-              'projectpath': u'D:/PyIron/PyIron_data/projects',
-              'status': u'finished',
-              'timestart': datetime.datetime(2016, 6, 24, 10, 17, 2, 429000),
-              'timestop': datetime.datetime(2016, 6, 24, 10, 17, 2, 463000),
-              'totalcputime': 0.034,
-              'username': u'test'},.......]
-
-    """
-    dict_clause = {}
-    # FOR GET_ITEMS_SQL: clause = []
-    if user is not None:
-        dict_clause["username"] = str(user)
-        # FOR GET_ITEMS_SQL: clause.append("username = '" + self.user + "'")
-    if sql_query is not None:
-        # FOR GET_ITEMS_SQL: clause.append(self.sql_query)
-        if "AND" in sql_query:
-            cl_split = sql_query.split(" AND ")
-        elif "and" in sql_query:
-            cl_split = sql_query.split(" and ")
-        else:
-            cl_split = [sql_query]
-        dict_clause.update(
-            {str(element.split()[0]): element.split()[2] for element in cl_split}
-        )
-    if job is not None:
-        dict_clause["job"] = str(job)
-
-    if project_path == "./":
-        project_path = ""
-    if recursive:
-        dict_clause["project"] = str(project_path) + "%"
-    else:
-        dict_clause["project"] = str(project_path)
-    if sub_job_name is None:
-        dict_clause["subjob"] = None
-    elif sub_job_name != "%":
-        dict_clause["subjob"] = str(sub_job_name)
-    if element_lst is not None:
-        dict_clause["element_lst"] = element_lst
-
-    s.logger.debug("sql_query: %s", str(dict_clause))
-    return database.get_items_dict(dict_clause)
-
 
 def get_db_columns(database):
     """
@@ -182,49 +92,16 @@ def job_table(
         pandas.Dataframe: Return the result as a pandas.Dataframe object
     """
     if not isinstance(database, FileTable):
-        if columns is None:
-            columns = ["job", "project", "chemicalformula"]
-        all_db = [
-            "id",
-            "status",
-            "chemicalformula",
-            "job",
-            "subjob",
-            "projectpath",
-            "project",
-            "timestart",
-            "timestop",
-            "totalcputime",
-            "computer",
-            "hamilton",
-            "hamversion",
-            "parentid",
-            "masterid",
-        ]
-
-        if all_columns:
-            columns = all_db
-        job_dict = _job_dict(
-            database=database,
-            sql_query=sql_query,
-            user=user,
-            project_path=project_path,
+        return database.job_table(
+            project=project_path,
             recursive=recursive,
-            element_lst=element_lst,
+            columns=columns,
+            all_columns=all_columns,
+            sort_by=sort_by,
+            max_colwidth=200,
+            full_table=full_table,
+            job_name_contains=job_name_contains
         )
-        if full_table:
-            pandas.set_option('display.max_rows', None)
-            pandas.set_option('display.max_columns', None)
-        else:
-            pandas.reset_option('display.max_rows')
-            pandas.reset_option('display.max_columns')
-        pandas.set_option("display.max_colwidth", max_colwidth)
-        df = pandas.DataFrame(job_dict, columns=columns)
-        if job_name_contains != '':
-            df = df[df.job.str.contains(job_name_contains)]
-        if sort_by in columns:
-            return df[columns].sort_values(by=sort_by)
-        return df[columns]
     else:
         database.update()
         return database.job_table(
@@ -364,8 +241,7 @@ def get_job_id(database, sql_query, user, project_path, job_specifier):
             return job_specifier  # is id
 
         job_specifier.replace(".", "_")
-        job_dict = _job_dict(
-            database=database,
+        job_dict = database._job_dict(
             sql_query=sql_query,
             user=user,
             project_path=project_path,
@@ -373,8 +249,7 @@ def get_job_id(database, sql_query, user, project_path, job_specifier):
             job=job_specifier
         )
         if len(job_dict) == 0:
-            job_dict = _job_dict(
-                database=database,
+            job_dict = database._job_dict(
                 sql_query=sql_query,
                 user=user,
                 project_path=project_path,
